Preliminary inspection through the BIOS
My inital diagnosis is that the SSD  is corrupted not the partitition the OS was on. It is likely that a sata cabled was moved or damaged becuase the kernel in memory incrementally lost access to hard drive when it was slowing losing functionality to access the keys. This conclusion is the same if the SSD started to fail